+++ flash/core/utilities/types.py
@@ -3,7 +3,8 @@ from typing import Any, Callable, Dict, Mapping, Optional, Sequence, Tuple, Unio
 from torch import nn
 from torchmetrics import Metric
 
-from flash.core.data.process import Deserializer, Postprocess, Preprocess, Serializer
+from flash.core.data.io.output import Output
+from flash.core.data.process import Deserializer, Postprocess, Preprocess
 
 MODEL_TYPE = Optional[nn.Module]
 LOSS_FN_TYPE = Optional[Union[Callable, Mapping, Sequence]]
@@ -15,4 +16,4 @@ METRICS_TYPE = Union[Metric, Mapping, Sequence, None]
 DESERIALIZER_TYPE = Optional[Union[Deserializer, Mapping[str, Deserializer]]]
 PREPROCESS_TYPE = Optional[Preprocess]
 POSTPROCESS_TYPE = Optional[Postprocess]
-SERIALIZER_TYPE = Optional[Union[Serializer, Mapping[str, Serializer]]]
+OUTPUT_TYPE = Optional[Union[Output, Mapping[str, Output]]]